Fixed circular import for spatial_from_spinorb function.
